--- docker-compose.yml
@@ -1,15 +1,0 @@
-version: '3.9'
-
-services:
-  app:
-    container_name: app
-    image: app
-    restart: always
-    build:
-      context: .
-      dockerfile: Dockerfile
-      target: base
-      args:
-        NODE_ENV: development
-    ports:
-      - '${PORT}:${PORT}'
