+++ src/locales/en-US/global.json
@@ -22,10 +22,11 @@
   "songIsBeingPlayed": "Track is being played.",
   "volume": "volume",
   "provider": "provider",
+  "pauseCapitalise": "Pause",
   "providerDesc": "Searches for tracks on the selected $t(global:provider).",
   "noMusicCurrentlyPlaying": "No track currently playing.",
   "volumeAlreadyTheSame": "The player is already set to this volume level.",
-  "volumeHasBeenModifiedTo": "The volume has been modified to ** {{ volume }} **/**100**%.",
+  "volumeHasBeenModifiedTo": "The volume has been modified to **{{ volume }}%**.",
   "somethingWentWrong": "Something went wrong. Please try again later.",
   "stop": "stop",
   "input": "input",
